test(api): extract worker loop fakes into shared module
